requirements.txt
flask
flask-cors
google-genai
python-dotenv